--- a/src/main/java/com/bootcamptoprod/service/MongoServiceClient.java
+++ b/src/main/java/com/bootcamptoprod/service/MongoServiceClient.java
@@ -28,6 +28,13 @@ public class MongoServiceClient {
         this.mongoClient = MongoClients.create(mongoUri);
     }
 
+    /**
+     * Get the MongoDB client for advanced operations.
+     */
+    public MongoClient getMongoClient() {
+        return this.mongoClient;
+    }
+
     // ========== DATABASE OPERATIONS ==========
 
     /**