--- a/src/main/java/com/deepai/SpringBootAiMongoMcpServerApplication.java
+++ b/src/main/java/com/deepai/SpringBootAiMongoMcpServerApplication.java
@@ -1,0 +1,13 @@
+package com.deepai;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBootAiMongoMcpServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringBootAiMongoMcpServerApplication.class, args);
+	}
+
+}