… .mcp.json
… .vscode/CodeCitations.md
… .vscode/mcp.json
… .vscode/tasks.json
… MongoDB-MCP-API-Testing-Guide.md
… MongoDB-MCP-Communication-Documentation.md
… MongoDB-MCP-Server-Postman-Collection.json
… MongoMcpApiController-API-Reference.md
… ONE-CLICK-TESTING-README.md
… PRODUCTION_READY_SUMMARY.md
… README.md
… TEST_REPORT.md
… automation-config.json
… docker-compose.yml
… logs/spring-boot-ai-mongo-mcp-server.log.2025-08-29.0.gz
… pom.xml
… run-api-tests.ps1
… src/main/java/com/bootcamptoprod/SpringBootAiMongoMcpServerApplication.java
… src/main/java/com/bootcamptoprod/controller/TestController.java
… src/main/java/com/bootcamptoprod/service/MongoAIService.java
… src/main/java/com/bootcamptoprod/service/MongoAdvancedAnalyticsService.java
… src/main/java/com/deepai/config/SecurityConfig.java
… src/main/java/com/deepai/controller/MongoMcpApiController.java
… src/main/java/com/deepai/service/MongoAIService.java
… src/main/java/com/deepai/service/MongoAdvancedAnalyticsService.java
… src/main/java/com/deepai/service/MongoServiceClient.java
… src/main/resources/application.properties
… src/main/resources/application.yml
… src/test/java/com/deepai/AdvancedMongoToolsBusinessLogicTest.java
… src/test/java/com/deepai/AllMongoToolsValidationTest.java